requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
clean-text==0.6.0
emoji==1.7.0
ftfy==6.1.1
idna==3.4
pyaes==1.6.1
pyasn1==0.5.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
rsa==4.9
Telethon==1.28.5
Unidecode==1.3.6
urllib3==2.0.2
wcwidth==0.2.6
